+++ CHANGELOG.md
@@ -11,3 +11,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - CrystalMap properties allow arrays with number of dimensions greater than 2
 - ANG file reader now recognises phase IDs defined in the header
+- EMsoft file reader reads unrefined Euler angles correctly